+++ apps/dashboard/src/lib/features/course/components/lesson/index.ts
@@ -10,5 +10,4 @@ export { default as Document } from './document/document.svelte';
 export { default as AddVideoModal } from './video/add-video-modal.svelte';
 export { default as AddDocumentModal } from './document/add-document-modal.svelte';
 export { default as LessonSettingsTab } from './lesson-settings-tab.svelte';
-export { default as LessonSummarizeButton } from './lesson-summarize-button.svelte';
 export { default as LessonMaterialActions } from './lesson-material-actions.svelte';
