+++ apps/dashboard/src/lib/features/course/components/lesson/lesson-material-actions.svelte
@@ -1,22 +1,36 @@
 <script lang="ts">
   import { Button } from '@cio/ui/base/button';
   import { Separator } from '@cio/ui/base/separator';
-  import { ContentIcon, HoverableItem } from '@cio/ui/custom/moving-icons';
+  import { ContentIcon, HoverableItem, PremiumIcon } from '@cio/ui/custom/moving-icons';
+  import { sendPromptToAssistant } from '$features/ai-assistant/utils/store';
   import { t } from '$lib/utils/functions/translations';
-  import LessonSummarizeButton from './lesson-summarize-button.svelte';
+  import { buildLessonSummarizePrompt } from './utils/lesson-summarize-prompt';
 
   interface Props {
     showTranscript?: boolean;
     showSummarize?: boolean;
+    showSeparators?: boolean;
     lessonId?: string;
     onTranscript?: () => void;
   }
 
-  let { showTranscript = false, showSummarize = false, lessonId = '', onTranscript = () => {} }: Props = $props();
+  let {
+    showTranscript = false,
+    showSummarize = false,
+    showSeparators = true,
+    lessonId = '',
+    onTranscript = () => {}
+  }: Props = $props();
+
+  function handleSummarizeLesson() {
+    sendPromptToAssistant(buildLessonSummarizePrompt(lessonId));
+  }
 </script>
 
 {#if showTranscript || showSummarize}
-  <Separator class="my-4" />
+  {#if showSeparators}
+    <Separator class="my-4" />
+  {/if}
 
   <div class="flex flex-wrap items-center gap-x-3 gap-y-2 py-1">
     {#if showTranscript}
@@ -35,9 +49,18 @@
     {/if}
 
     {#if showSummarize}
-      <LessonSummarizeButton {lessonId} />
+      <HoverableItem>
+        {#snippet children(isHovered)}
+          <Button variant="ghost" size="sm" class="ui:px-2" onclick={handleSummarizeLesson}>
+            <PremiumIcon {isHovered} size={16} />
+            {$t('course.navItem.lessons.materials.summarize_lesson')}
+          </Button>
+        {/snippet}
+      </HoverableItem>
     {/if}
   </div>
 
-  <Separator class="my-4" />
+  {#if showSeparators}
+    <Separator class="my-4" />
+  {/if}
 {/if}
